data: move conversion data into lenskit.data.sources
