+++ src/lenskit/cli/data/describe.py
@@ -12,10 +12,8 @@ import click
 from rich.console import Console
 from rich.markdown import Markdown
 
-from lenskit.data import Dataset
+from lenskit.data import Dataset, load_amazon_ratings, load_movielens
 from lenskit.data._summary import save_stats
-from lenskit.data.amazon import load_amazon_ratings
-from lenskit.data.movielens import load_movielens
 from lenskit.logging import get_logger
 
 _log = get_logger(__name__)
