+++ tests/data/test_huge.py
@@ -17,8 +17,7 @@ import pyarrow as pa
 
 from pytest import mark
 
-from lenskit.data import DatasetBuilder, from_interactions_df
-from lenskit.data.movielens import load_movielens_df
+from lenskit.data import DatasetBuilder, from_interactions_df, load_movielens_df
 from lenskit.logging import get_logger
 
 pytestmark = mark.skipif("LK_HUGE_TEST" not in os.environ, reason="huge tests disabled")
